requirements.txt
bidict==0.21.4
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
dnspython==2.2.1
eventlet==0.33.1
Flask==2.0.3
Flask-SocketIO==5.3.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
zipp==3.6.0
